requirements.txt
torch
torchvision
opencv-python
scikit-learn
onnxruntime
numpy
pandas
matplotlib
pillow